chore: override page title & description on sub pages
