+++ apps/web/src/app/[lang]/app/[type]/crm/[data]/[id]/subcompany/new/layout.tsx
@@ -1,6 +1,4 @@
 "use client";
-import Link from "next/link";
-import { PageHeader } from "@repo/ayasofyazilim-ui/molecules/page-header";
 import { Card, CardContent } from "@/components/ui/card";
 import { getResourceDataClient } from "src/language-data/CRMService";
 import { useLocale } from "src/providers/locale";
@@ -19,18 +17,28 @@ export default function Layout({ children, params }: LayoutProps) {
   const { resources } = useLocale();
   const languageData = getResourceDataClient(resources, params.lang);
 
+  const pageBack = getBaseLink(
+    `/crm/${params.data}/${params.id}`,
+    true,
+    params.lang,
+    true,
+    "admin",
+  );
   return (
     <>
-      <PageHeader
-        LinkElement={Link}
-        description={languageData["SubCompany.Description"]}
-        href={getBaseLink(`/app/admin/crm/${params.data}/${params.id}`)}
-        title={languageData["SubCompany.Description"]}
-      />
       <div className="flex h-full w-full flex-row">
         <Card className="m-0 w-full overflow-auto border-0 bg-transparent bg-white pt-5 shadow-none">
           <CardContent>{children}</CardContent>
         </Card>
+      </div>
+      <div className="hidden" id="page-title">
+        {languageData["SubCompany.New"]}
+      </div>
+      <div className="hidden" id="page-description">
+        {languageData["SubCompany.New"]}
+      </div>
+      <div className="hidden" id="page-back-link">
+        {pageBack}
       </div>
     </>
   );
